Added functionality to the search bar and button
Both key presses and clicking the search button now trigger a search. The search results are displayed in the same fashion as before the search: ten students per page and clickable buttons to select different pages of ten. It is worth noting that clearing the search bar causes all students to be displayed as they were before a search was triggered.